requirements.txt
requests
pandas
pyarrow
fastparquet
apache-airflow
